requirements.txt
# Core web framework
flask>=3.0.0
flask-sqlalchemy>=3.1.0

# Data & market data
pandas>=2.0.0
yfinance>=0.2.40
pandas-market-calendars>=4.3.0

# Environment & config
python-dotenv>=1.0.0

# Timezone
pytz>=2024.1

# Telegram bot (polling mode)
python-telegram-bot>=20.0,<21.0

# Claude / Anthropic API
anthropic>=0.34.0

# News headlines
newsapi-python>=0.2.7

# WSGI server for production
gunicorn>=22.0.0
